Vendor CAME rather than just implement extensions.
