+++ modules/util/create.py
@@ -68,10 +68,10 @@ from modules.util.enum.NoiseScheduler import NoiseScheduler
 from modules.util.enum.Optimizer import Optimizer
 from modules.util.enum.TrainingMethod import TrainingMethod
 from modules.util.lr_scheduler_util import *
+from modules.util.optimizer import CAME
 from modules.util.optimizer.adafactor_extensions import step_adafactor, patch_adafactor
 from modules.util.optimizer.adam_extensions import step_adam, patch_adam
 from modules.util.optimizer.adamw_extensions import step_adamw, patch_adamw
-from modules.util.optimizer.came_extensions import patch_came
 
 
 def create_model_loader(
@@ -646,9 +646,7 @@ def create_optimizer(
 
         # CAME Optimizer
         case Optimizer.CAME:
-            import came_pytorch
-
-            optimizer = came_pytorch.CAME(
+            optimizer = CAME.CAME(
                 params=parameters,
                 lr=config.learning_rate,
                 eps=(optimizer_config.eps if optimizer_config.eps is not None else 1e-30,
@@ -657,9 +655,8 @@ def create_optimizer(
                        optimizer_config.beta2 if optimizer_config.beta2 is not None else 0.999,
                        optimizer_config.beta3 if optimizer_config.beta3 is not None else 0.9999),
                 weight_decay=optimizer_config.weight_decay if optimizer_config.weight_decay is not None else 0,
+                stochastic_rounding=optimizer_config.stochastic_rounding
             )
-
-            patch_came(optimizer, optimizer_config.stochastic_rounding)
 
     if state_dict is not None:
         for i, params in enumerate(parameters):
