+++ requirements-global.txt
@@ -30,7 +30,6 @@ open-clip-torch==2.23.0
 dadaptation==3.2 # dadaptation optimizers
 lion-pytorch==0.1.2 # lion optimizer
 prodigyopt==1.0 # prodigy optimizer
-came-pytorch==0.1.3  # CAME optimizer
 
 # ui
 customtkinter==5.2.1
